fix: rename blogpost and newsarticle component to keep naming convention for Storyblok blok components
